fix: resolve integer out of range errors for NSE data tables
- Change `change_in_oi` on `bhavcopy_fo` table to `BigInteger` because high volume days result in `change_in_oi` exceeding Postgres 32-bit Integer maximums (2.14B).
- Upgrade `total_traded_qty`, `total_trades`, `open_interest`, and `change_in_oi` on `bhavcopy` legacy table to `BigInteger` to prevent `NumericValueOutOfRange` during import.
- Upgrade contract metrics on `fao_participant_oi` from `Integer` to `BigInteger`.

+++ backend/domain/market/models.py
@@ -52,11 +52,12 @@ class Bhavcopy(Base):
     settlement_price = Column(Float)  # SttlmPric
 
     # Volume & OI fields
-    total_traded_qty = Column(Integer) # TtlTradgVol
+    from sqlalchemy import BigInteger
+    total_traded_qty = Column(BigInteger) # TtlTradgVol
     total_traded_val = Column(Float) # TtlTrfVal
-    total_trades = Column(Integer) # TtlNbOfTxsExctd
-    open_interest = Column(Integer)  # OpnIntrst
-    change_in_oi = Column(Integer)   # ChngInOpnIntrst
+    total_trades = Column(BigInteger) # TtlNbOfTxsExctd
+    open_interest = Column(BigInteger)  # OpnIntrst
+    change_in_oi = Column(BigInteger)   # ChngInOpnIntrst
 
     lot_size = Column(Integer) # NewBrdLotQty
